Support terminals with light background
In terminals with a light background, white text is very hard or even impossible to read.
This revision changes white text to "reset", which is a pseudo-color that resets to the default colors.

An alternative approach could be to use https://github.com/Canop/terminal-light to check the terminal background luma and determine a suitable default text color from that,
but since we're not styling the terminal background the reset approach seemed more straightforward.